+++ morphir/runtime/src/org/finos/morphir/runtime/quick/QuickMorphirRuntime.scala
@@ -9,6 +9,7 @@ import org.finos.morphir.ir.distribution.Distribution
 import org.finos.morphir.ir.Type.Type
 import org.finos.morphir.runtime.*
 import org.finos.morphir.runtime.exports.*
+import org.finos.morphir.runtime.quick.testing.UnitTesting
 import org.finos.morphir.runtime.services.sdk.MorphirSdk
 import org.finos.morphir.runtime.Utils.*
 import org.finos.morphir.ir.conversion.*
